Analysis feature brokerclient (#272)
* Added Broker client to analysis engine

* Added interfaces for BrokerClient - Kyle Barry (u19232510)

* Added resource classes to BrokerClient - Kyle Barry (u19232510)

* Added ServerInformation class - Kyle Barry (u19232510)

* Server information is collected and sent to Kafka - Kyle Barry (u19232510)

* Adding Command handler objects for Broker Client - Kyle Barry (u19232510)

* Added analysis command - Kyle Barry (u19232510)

* Updated AnalysisCommand - Kyle Barry (u19232510)

* Refactored code in the BrokerClient - Kyle Barry (u19232510)

* BrokerClient registers with Broker and sends performance information to Broker. - Kyle Barry (u19232510)

* Added extra touches to Broker. - Kyle Barry (u19232510)

* Added fixes to Broker. - Kyle Barry (u19232510)

* Added fixes to Broker. - Kyle Barry (u19232510)

* Added support for heartbeats to Broker. - Kyle Barry (u19232510)

* Get UserId from AnalysisRequest in Broker. - Kyle Barry (u19232510)

* Broker sends live stream information to all connections with the same user id. - Kyle Barry (u19232510)

* Updated WebsocketConnection to handle client IDs being sent first before analysis request - Kyle Barry (u19232510)

* Broker is now fully implemented - Kyle Barry (u19232510)

* Broker is now fully implemented - Kyle Barry (u19232510)

* Created lock for creating and deleting topics via Process. - Kyle Barry (u19232510)

* Created lock for creating and deleting topics via Process. - Kyle Barry (u19232510)

* Created lock for creating and deleting topics via Process. - Kyle Barry (u19232510)

* Created single Thread pool. - Kyle Barry (u19232510)

* Broker sends streamId to clients. - Kyle Barry (u19232510)

* Broker ensures request comes from correct backend. - Kyle Barry (u19232510)

* Broker disconnects when it catches errors and is now fully working. - Kyle Barry (u19232510)

* Copied updated Analysis engine and Broker can synchronise with client when client asks for synchronisation. - Kyle Barry (u19232510)

* Updated synchronisation logic. - Kyle Barry (u19232510)

* Updated synchronisation logic. - Kyle Barry (u19232510)

* Fixed concurrent modification issue when removing broken connections. - Kyle Barry (u19232510)

+++ broker/BrokerServer/src/main/java/managers/topicmanager/TopicManager.java
@@ -3,6 +3,7 @@ package managers.topicmanager;
 import com.google.gson.*;
 import com.google.gson.reflect.TypeToken;
 import logger.EventLogger;
+import managers.concurrencymanager.ConcurrencyManager;
 import org.apache.commons.io.IOUtil;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 
@@ -12,14 +13,15 @@ import java.net.URL;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
-import java.util.concurrent.*;
 import java.util.concurrent.locks.ReentrantLock;
 
 public class TopicManager {
     private static TopicManager _instance;
     private final ReentrantLock lock = new ReentrantLock();
     private final ReentrantLock responseLock = new ReentrantLock();
-    private final Executor actionExecutor = Executors.newFixedThreadPool(2);
+    private final ConcurrencyManager actionExecutor = ConcurrencyManager.getInstance();
+    private final ReentrantLock topicLock = new ReentrantLock();
+    private volatile boolean isLocked = false;
     private Map<String, String> responseIds;
 
     private TopicManager() {
@@ -52,13 +54,20 @@ public class TopicManager {
         lock.lock();
         try {
             actionExecutor.execute(() -> {
+                if (!isLocked) {
+                    EventLogger.getLogger().warn("Not deleting topic as lock was not acquired");
+                    return;
+                }
                 String exec = System.getenv("KAFKA_DELETE_TOPIC").replace("{topic}", topic);
                 try {
                     ProcessBuilder builder = new ProcessBuilder(exec.split(" "));
-                    builder.start();
+                    Process proc = builder.start();
                     deleteServer(topic);
-                } catch (IOException e) {
+                    proc.waitFor();
+                } catch (IOException | InterruptedException e) {
                     EventLogger.getLogger().error(e.getMessage());
+                } finally {
+                    isLocked = false;
                 }
             });
         }finally {
@@ -75,6 +84,10 @@ public class TopicManager {
         lock.lock();
         try {
             actionExecutor.execute(() -> {
+                if (!isLocked) {
+                    EventLogger.getLogger().warn("Not adding topic as lock was not acquired");
+                    return;
+                }
                 EventLogger.getLogger().info("Creating topic " + topic);
                 ProcessBuilder builder = new ProcessBuilder(System.getenv("KAFKA_CREATE_TOPIC").
                         replace("{topic}", topic).split(" "));
@@ -83,6 +96,8 @@ public class TopicManager {
                     proc.waitFor();
                 } catch (IOException | InterruptedException e) {
                     EventLogger.getLogger().logException(e);
+                } finally {
+                    isLocked = false;
                 }
             });
         }finally {
@@ -180,5 +195,22 @@ public class TopicManager {
         StringWriter writer = new StringWriter();
         IOUtil.copy(is, writer, "UTF-8");
         return writer;
+    }
+
+    public void lockTopic() {
+        topicLock.lock();
+        isLocked = true;
+    }
+
+    public void unlockTopic() {
+
+        while(isLocked) {
+            try {
+                Thread.sleep(200);
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
+        topicLock.unlock();
     }
 }
